% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{Brazil_films_df}
\alias{Brazil_films_df}
\title{Films Shown at Brazilian Film Festivals (2007–2011)}
\format{
A data frame with 25 observations and 6 variables:
\describe{
  \item{year}{Year of the film festival (integer)}
  \item{regE}{Festival region (factor with 5 levels)}
  \item{F}{Number of films shown (integer)}
  \item{D}{Number of directors (integer)}
  \item{MD}{Number of male directors (integer)}
  \item{WD}{Number of female directors (integer)}
}
}
\source{
Data taken from the bpca package version 1.3-6
}
\usage{
data(Brazil_films_df)
}
\description{
This dataset, Brazil_films_df, is a data frame containing information on films shown at five
different film festivals in Brazil from 2007 to 2011. The dataset includes 25 observations and 6 variables,
summarizing the number of films, directors, male and female directors, and regional categories for each year.
}
\details{
The dataset name has been kept as 'Brazil_films_df' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
BrazilDataAPI package and assists users in identifying its specific characteristics.
The suffix 'df' indicates that the dataset is a data frame. The original content has not been modified
in any way.
}
