% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazildataapi-package.R
\docType{package}
\name{BrazilDataAPI}
\alias{BrazilDataAPI}
\alias{BrazilDataAPI-package}
\title{BrazilDataAPI: Access Brazilian Data via APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'BrasilAPI', 'REST Countries API','Nager.Date API', and 'World Bank API', related to Brazil's postal codes, banks, holidays,company registrations, international country indicators, public holidays information, and economic development data. Additionally, the package includes curated datasets related to Brazil, covering topics such as demographic data,river levels, environmental emission factors, film festivals, and yellow fever outbreak records.
}
}
\details{
BrazilDataAPI: Access Brazilian Data via APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Brazilian Data via APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/brazildataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
