% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_brazil_rates.R
\name{get_brazil_rates}
\alias{get_brazil_rates}
\title{Get Official Interest Rates and Indexes from Brazil}
\usage{
get_brazil_rates()
}
\value{
A tibble with the following columns:
\itemize{
  \item \code{nome}: Name or acronym of the rate/index.
  \item \code{valor}: Current value of the rate/index.
}
}
\description{
This function retrieves official interest rates and indexes from the BrazilAPI
endpoint: `https://brasilapi.com.br/api/taxas/v1`.
}
\examples{
\dontrun{
taxas <- get_brazil_rates()
print(taxas)
}

}
\seealso{
\code{\link[httr]{GET}}, \code{\link[jsonlite]{fromJSON}}, \code{\link[dplyr]{as_tibble}}
}
