% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\name{manaus_ts}
\alias{manaus_ts}
\title{Monthly Average Heights of the Rio Negro at Manaus (1903–1992)}
\format{
A univariate time series of class \code{ts} with 1080 monthly observations from 1903 to 1992:
\describe{
  \item{values}{Monthly average river heights (numeric)}
}
}
\source{
Data taken from the boot package version 1.3-31
}
\usage{
data(manaus_ts)
}
\description{
This dataset, manaus_ts, is a univariate time series of monthly average river heights of the Rio Negro at Manaus.
The series contains 1080 observations spanning 90 years, from January 1903 to December 1992. Each value represents
the monthly average of the daily stages (heights) of the Rio Negro. Manaus is located 18 km upstream from the confluence
of the Rio Negro with the Amazon River, and due to the minimal slope and flatland affluents, the measurements can be considered
a good approximation of the water level at the confluence.
}
\details{
The dataset name has been kept as 'manaus_ts' to avoid confusion with other datasets
in the R ecosystem. This naming convention helps distinguish this dataset as part of the
BrazilDataAPI package and assists users in identifying its specific characteristics.
The suffix 'ts' indicates that the dataset is a time series object. The original content has not been modified
in any way.
}
