% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazildataapi-package.R
\docType{package}
\name{BrazilDataAPI}
\alias{BrazilDataAPI}
\alias{BrazilDataAPI-package}
\title{BrazilDataAPI: Access Brazilian Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets}
\description{
{
This package provides functions to access data from the 'BrasilAPI' and the 'REST Countries API',related to Brazil's postal codes, banks, holidays, company registrations, and international country indicators.
}
}
\details{
BrazilDataAPI: Access Brazilian Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Access Brazilian Economic, Demographic, Environmental, and Geopolitical Data via RESTful APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/brazildataapi}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
